Enable mouse wheel in freehand draw mode
Previously, events weren't being propagated if the active draw
interaction had its `freehand` property set to true. This means that
you can't use your mouse wheel to zoom the map in freehand mode.

This change sets the default `pass` value to true and adds additional
logic to determine when it should be set to false (and stop event
propagation).